Use relative paths instead of absolute for symlink
Use relative paths instead of absolute paths for creating the
symbolic link in order to allow Windows to see the symbolic
link created in linux.

This DS allows to create a symbolic link to a specific folder.

In the specific case of Mistral, this DS should allow to
store the data from XMController in different folders, by
changing the symbolic link. XMController will store all
its data in a same folder that will point to different user
folders. The symbolic link will be set to point from the
same folder to the different user folders.